+++ expenses_backend/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for expenses_site project.
+WSGI config for expenses_backend project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "expenses_site.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "expenses_backend.settings")
 
 application = get_wsgi_application()